pagerank double

+++ src/edu/nyu/cs/cs2580/CorpusAnalyzerPagerank.java
@@ -219,8 +219,8 @@ public class CorpusAnalyzerPagerank extends CorpusAnalyzer implements Serializab
     return null;
   }
 
-  public Integer getPagerank(String doc) {
-    return (int) (_ranked_docs.containsKey(doc) ? _ranked_docs.get(doc) : 0);
+  public double getPagerank(String doc) {
+    return  (_ranked_docs.containsKey(doc) ? _ranked_docs.get(doc) : 0.0);
   }
 
 }
